requirements.txt
dwave-ocean-sdk
numpy
